requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
Django==1.10.6
django-allauth==0.31.0
django-cors-headers==2.0.2
django-filter==1.0.1
django-rest-auth==0.9.1
django-taggit==0.22.0
djangorestframework==3.6.2
oauthlib==2.0.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.7.1
pyparsing==2.2.0
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
